good performance

+++ vllm/model_executor/layers/fused_moe/fused_moe.py
@@ -385,10 +385,10 @@ def fused_moe(
 
     intermediate_cache1 = torch.empty((M, topk_ids.shape[1], N),
                                       device=hidden_states.device,
-                                      dtype=torch.float8_e4m3fn)
+                                      dtype=torch.float16)
     intermediate_cache2 = torch.empty((M * topk_ids.shape[1], N // 2),
                                       device=hidden_states.device,
-                                      dtype=torch.float8_e4m3fn)
+                                      dtype=torch.float16)
     intermediate_cache3 = torch.empty((M, topk_ids.shape[1], w2.shape[1]),
                                       device=hidden_states.device,
                                       dtype=torch.float16)
@@ -406,7 +406,8 @@ def fused_moe(
 
     ops.scaled_silu_and_mul(intermediate_cache2, intermediate_cache1.view(-1, N), s2)
 
-    invoke_fused_moe_kernel(intermediate_cache2, w2, intermediate_cache3,
+    invoke_fused_moe_kernel(intermediate_cache2.to(dtype=torch.float8_e4m3fn),
+                            w2, intermediate_cache3,
                             topk_weights, topk_ids, sorted_token_ids,
                             expert_ids, num_tokens_post_padded, True, 1,
                             config, compute_type=tl.float16)
